telegram_agent/requirements.txt
anthropic>=0.40.0
python-telegram-bot>=21.0
duckduckgo-search>=6.0
python-dotenv>=1.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
aiofiles>=23.0.0
